Separate shape logic
