+++ apps/desktop/src/global/classes/canvasObjects/MarcherShape.ts
@@ -1,14 +1,11 @@
 import OpenMarchCanvas from "./OpenMarchCanvas";
 import type { ShapePage } from "electron/database/tables/ShapePageTable";
-import {
-    ShapePath,
-    ShapePoint,
-    StaticMarcherShape,
-    SvgCommandEnum,
-    VanillaPoint,
-} from "./StaticMarcherShape";
 import MarcherPage from "@/global/classes/MarcherPage";
 import type { ModifiedShapePageMarcherArgs } from "electron/database/tables/ShapePageMarcherTable";
+import { ShapePath } from "./ShapePath";
+import { ShapePoint } from "./ShapePoint";
+import { StaticMarcherShape, VanillaPoint } from "./StaticMarcherShape";
+import { SvgCommandEnum } from "./SvgCommand";
 
 /**
  * A MarcherShape is StaticMarcherShape that is stored in the database and updates the database as it is modified.
