+++ apps/desktop/src/global/classes/canvasObjects/StaticMarcherShape.ts
@@ -2,7 +2,10 @@ import { fabric } from "fabric";
 import OpenMarchCanvas from "./OpenMarchCanvas";
 import MarcherPage from "../MarcherPage";
 import CanvasMarcher from "./CanvasMarcher";
-import { DEFAULT_FIELD_THEME, FieldTheme, rgbaToString } from "../FieldTheme";
+import { ShapePoint } from "./ShapePoint";
+import { Coordinate, SvgCommandEnum } from "./SvgCommand";
+import { ShapePath } from "./ShapePath";
+import { ShapePointController } from "./ShapePointController";
 
 /**
  * An SVG point in the StaticMarcherShape path.
@@ -10,14 +13,6 @@ import { DEFAULT_FIELD_THEME, FieldTheme, rgbaToString } from "../FieldTheme";
  * This is the lowest type of the point and is how a path's point is represented in fabric.js
  */
 export type VanillaPoint = [string, ...number[]];
-
-export enum SvgCommandEnum {
-    MOVE = "M",
-    LINE = "L",
-    QUADRATIC = "Q",
-    CUBIC = "C",
-    CLOSE = "Z",
-}
 
 /**
  * A canvas path with control points.
@@ -513,652 +508,6 @@ export class StaticMarcherShape {
             this.disableControl();
         }
     }
-}
-
-/**
- * Represents a control point for a StaticMarcherShape object, which is part of a path in a canvas.
- * The control point is associated with a specific point and coordinate index in the path.
- * It handles the movement and modification of the control point, and updates the corresponding
- * coordinates in the path.
- */
-class ShapePointController extends fabric.Circle {
-    /**
-     * The path this control point is a part of
-     */
-    marcherShape: StaticMarcherShape;
-    /**
-     * The index of the point in the path.
-     * I.e. `this.marcherShape.path.path[this.pathIndex]` is the point this control point is associated with
-     */
-    pointIndex: number;
-    /**
-     * The index of the coordinates in the point.
-     * I.e. `this.marcherShape.path.path[this.pointIndex][this.coordIndex]` is the coordinate this control point is associated with
-     */
-    coordIndex: number;
-
-    /** The point directly before this one in the path */
-    incomingPoint: ShapePointController | null;
-    /** The point directly after this one in the path */
-    outgoingPoint: ShapePointController | null;
-    /** A fabric line that leads to the outgoing point to visual the relationship */
-    outgoingLine: fabric.Line | null;
-
-    canvas?: OpenMarchCanvas;
-    left: number;
-    top: number;
-
-    constructor({
-        marcherShape,
-        pointIndex,
-        coordIndex,
-        canvas,
-        incomingPoint = null,
-        outgoingPoint = null,
-    }: {
-        marcherShape: StaticMarcherShape;
-        pointIndex: number;
-        coordIndex: number;
-        canvas: OpenMarchCanvas;
-        incomingPoint?: ShapePointController | null;
-        outgoingPoint?: ShapePointController | null;
-    }) {
-        if (coordIndex === 0)
-            console.warn(
-                "The coordinate index in the SVG point likely should not be 0, as this is the SVG command",
-            );
-        if (!marcherShape.shapePath.path) {
-            throw new Error("The parent path does not have a path");
-        }
-
-        super({
-            strokeWidth: 4,
-            radius: 6,
-            fill: "#fff",
-            stroke: rgbaToString(canvas.fieldProperties.theme.shape),
-            hasBorders: false,
-            originX: "center",
-            originY: "center",
-            hasControls: false,
-            hoverCursor: "point",
-        });
-
-        this.marcherShape = marcherShape;
-        this.pointIndex = pointIndex;
-        this.coordIndex = coordIndex;
-
-        this.incomingPoint = incomingPoint;
-        // If the incoming point does not have this as an outgoing point, set it to this
-        if (this.incomingPoint && !(this.incomingPoint.outgoingPoint === this))
-            this.incomingPoint.outgoingPoint = this;
-
-        this.outgoingPoint = outgoingPoint;
-
-        this.outgoingLine = null;
-        this.canvas = canvas;
-
-        const point = this.getPathCoordinates();
-        if (!point) {
-            throw new Error("The point does not have coordinates");
-        }
-        this.left = point.left;
-        this.top = point.top;
-
-        // Create the new listeners
-        this.on("moving", this.moveHandler.bind(this));
-        this.on("modified", this.modifiedHandler.bind(this));
-    }
-
-    destroy() {
-        if (this.canvas) {
-            if (this.outgoingLine) this.canvas.remove(this.outgoingLine);
-            this.canvas.remove(this);
-        }
-    }
-
-    /**
-     * Refreshes the coordinates of the path point associated with this StaticMarcherShape object.
-     *
-     * This method updates the coordinates of the corresponding path point in the parent path's path array. It calculates the new coordinates based on the current position of the control point and the curve offset, and then sets the updated coordinates in the path array. Finally, it marks the parent path as dirty to trigger a redraw.
-     */
-    refreshParentPathCoordinates() {
-        if (!this.marcherShape.shapePath.path) {
-            console.error("The parent path does not have a path");
-            return;
-        }
-        if (!this.left || !this.top) {
-            console.error("The control point does not have coordinates");
-            return;
-        }
-        (
-            this.marcherShape.shapePath.path[
-                this.pointIndex
-            ] as unknown as number[]
-        )[this.coordIndex] =
-            this.left - this.marcherShape.moveOffset.fromInitial.x;
-        (
-            this.marcherShape.shapePath.path[
-                this.pointIndex
-            ] as unknown as number[]
-        )[this.coordIndex + 1] =
-            this.top - this.marcherShape.moveOffset.fromInitial.y;
-        this.marcherShape.shapePath.setCoords();
-        this.marcherShape.shapePath.dirty = true;
-    }
-
-    /**
-     * Gets the coordinates of the path point associated with this StaticMarcherShape object.
-     *
-     * This method retrieves the coordinates of the path point that corresponds to this StaticMarcherShape object. It extracts the left and top coordinates from the path array and returns them as an object.
-     *
-     * @returns {Object} An object containing the left and top coordinates of the path point, or `null` if the parent path does not have a path.
-     * @property {number} left - The left coordinate of the path point.
-     * @property {number} top - The top coordinate of the path point.
-     */
-    getPathCoordinates() {
-        if (!this.marcherShape.shapePath.path) {
-            console.error("The parent path does not have a path");
-            return;
-        }
-        const vanillaPoint = this.marcherShape.shapePath.path[
-            this.pointIndex
-        ] as unknown as number[];
-        const point = {
-            left: vanillaPoint[this.coordIndex],
-            top: vanillaPoint[this.coordIndex + 1],
-        };
-        return point;
-    }
-
-    /**
-     * Handles the movement of the parent path by updating the control point's coordinates
-     */
-    handleParentMove() {
-        const vanillaPoint = this.getPathCoordinates();
-        if (!vanillaPoint) {
-            throw new Error("The point does not have coordinates");
-        }
-        this.left =
-            vanillaPoint.left + this.marcherShape.moveOffset.fromInitial.x;
-        this.top =
-            vanillaPoint.top + this.marcherShape.moveOffset.fromInitial.y;
-
-        this.refreshLines();
-    }
-
-    /**
-     * Handles the movement of the control point by updating the path coordinates and refreshing the connected lines.
-     *
-     * This method is called when the control point is moved. It updates the coordinates of the corresponding path point, and then refreshes the incoming and outgoing lines connected to the control point.
-     *
-     * @param e - The fabric.js event object containing information about the move event.
-     */
-    moveHandler(e: fabric.IEvent) {
-        this.marcherShape.shapePath.objectCaching = false;
-        if (this.marcherShape.shapePath.path && this.left && this.top) {
-            this.refreshParentPathCoordinates();
-            this.marcherShape.distributeMarchers();
-            this.marcherShape.bringControlPointsToFront();
-        }
-        this.marcherShape.dirty = true;
-        this.refreshLines();
-    }
-
-    /**
-     * Redraws the parent path of the StaticMarcherShape object.
-     *
-     * This method is called when the StaticMarcherShape object has been modified, and the parent path needs to be redrawn.
-     *
-     * @param path - The path of the parent object, as an array of (string | number)[][] elements.
-     */
-    modifiedHandler(e: fabric.IEvent) {
-        if (!this.marcherShape.shapePath.path) {
-            console.error("The parent path does not have a path");
-            return;
-        }
-
-        this.marcherShape.recreatePath(
-            this.marcherShape.shapePath.path as unknown as VanillaPoint[],
-        );
-    }
-
-    /**
-     * Refreshes both the first point of the outgoing line and the second point of the incoming line.
-     *
-     * Each line is only updated if they exist.
-     */
-    refreshLines() {
-        if (this.incomingPoint) {
-            const incomingLine = this.incomingPoint.outgoingLine;
-            if (incomingLine) {
-                incomingLine.set({
-                    x2: this.left,
-                    y2: this.top,
-                });
-            }
-        }
-        if (this.outgoingLine) {
-            this.outgoingLine.set({
-                x1: this.left,
-                y1: this.top,
-            });
-        }
-    }
-
-    /**
-     * Draws a dashed line from this control point to the outgoing point
-     *
-     * @returns The line that was drawn
-     */
-    drawOutgoingLine(): fabric.Line | null {
-        if (!this.outgoingPoint) {
-            // There is no outgoing point, return null
-            return null;
-        }
-        if (!this.outgoingPoint.left || !this.outgoingPoint.top) {
-            console.error("The outgoing point has no coordinates");
-            return null;
-        }
-        const outgoingLine = new fabric.Line(
-            [
-                this.left,
-                this.top,
-                this.outgoingPoint.left,
-                this.outgoingPoint.top,
-            ],
-            {
-                stroke: rgbaToString(
-                    this.canvas?.fieldProperties.theme.shape ?? {
-                        r: 126,
-                        g: 34,
-                        b: 206,
-                        a: 1,
-                    },
-                ),
-                strokeWidth: 1,
-                strokeDashArray: [5, 5],
-                selectable: false,
-                evented: false,
-            },
-        );
-        this.outgoingLine = outgoingLine;
-        if (!this.canvas) {
-            console.error(
-                "The canvas is not defined. The outgoing lines will not be drawn",
-            );
-        } else this.canvas?.add(outgoingLine);
-
-        return outgoingLine;
-    }
-}
-
-/**
- * The fabric.Path object that represents the path of the shape.
- */
-export class ShapePath extends fabric.Path {
-    static fieldTheme: FieldTheme = DEFAULT_FIELD_THEME;
-
-    /**
-     * @param points The points to draw the path from
-     */
-    constructor(points: ShapePoint[]) {
-        super(ShapePoint.pointsToString(points), {
-            fill: "",
-            strokeWidth: 2,
-            stroke: rgbaToString(ShapePath.fieldTheme.shape),
-            objectCaching: true,
-            hasControls: false,
-            borderColor: "#0d6efd",
-            borderScaleFactor: 2,
-        });
-        this.disableControl();
-    }
-
-    enableControl() {
-        this.stroke = rgbaToString(ShapePath.fieldTheme.shape);
-        this.strokeWidth = 2;
-        this.strokeDashArray = [];
-        this.selectable = true;
-        this.backgroundColor = rgbaToString({
-            ...ShapePath.fieldTheme.tempPath,
-            a: 0.2,
-        });
-        this.hoverCursor = "move";
-    }
-
-    disableControl() {
-        this.stroke = rgbaToString(ShapePath.fieldTheme.tempPath);
-        this.strokeWidth = 1;
-        this.strokeDashArray = [5, 3];
-        this.selectable = false;
-        this.backgroundColor = "transparent";
-        this.hoverCursor = "default";
-    }
-
-    get points(): ShapePoint[] {
-        if (!this.path) {
-            console.error("The path is not defined");
-            return [];
-        }
-        return ShapePoint.fromArray(this.path as any as VanillaPoint[]);
-    }
-
-    /**
-     * String representation of all the points in the path.
-     *
-     * @returns A string representation of the path. E.g. "M 0 0 L 100 100"
-     */
-    toString() {
-        return this.points.map((point) => point.toString()).join(" ");
-    }
-
-    equals(other: ShapePath) {
-        return this.toString() === other.toString();
-    }
-}
-
-type Coordinate = {
-    x: number;
-    y: number;
-};
-
-interface SvgCommand {
-    readonly readableDescription: string;
-    readonly command: SvgCommandEnum;
-    readonly numberOfCoordinates: number;
-}
-
-export const SvgCommands: {
-    [key in SvgCommandEnum]: SvgCommand;
-} = {
-    [SvgCommandEnum.MOVE]: {
-        readableDescription: "Move",
-        command: SvgCommandEnum.MOVE,
-        numberOfCoordinates: 1,
-    },
-    [SvgCommandEnum.LINE]: {
-        readableDescription: "Line",
-        command: SvgCommandEnum.LINE,
-        numberOfCoordinates: 1,
-    },
-    [SvgCommandEnum.QUADRATIC]: {
-        readableDescription: "Single Curve",
-        command: SvgCommandEnum.QUADRATIC,
-        numberOfCoordinates: 2,
-    },
-    [SvgCommandEnum.CUBIC]: {
-        readableDescription: "Double Curve",
-        command: SvgCommandEnum.CUBIC,
-        numberOfCoordinates: 6,
-    },
-    [SvgCommandEnum.CLOSE]: {
-        readableDescription: "Close",
-        command: SvgCommandEnum.CLOSE,
-        numberOfCoordinates: 0,
-    },
-};
-
-export const secondSegmentSvgCommands: SvgCommand[] = [
-    SvgCommands[SvgCommandEnum.LINE],
-    SvgCommands[SvgCommandEnum.QUADRATIC],
-    SvgCommands[SvgCommandEnum.CUBIC],
-];
-/**
- * Represents a single point in a shape path, with a command and coordinates.
- * The `ShapePoint` class provides methods to work with and manipulate these points.
- */
-export class ShapePoint {
-    command: SvgCommandEnum;
-    coordinates: Coordinate[];
-
-    private constructor(command: SvgCommandEnum, coordinates: Coordinate[]) {
-        this.command = command;
-        this.coordinates = coordinates;
-    }
-
-    static checkCommand(command: string): SvgCommandEnum {
-        if (
-            !Object.values(SvgCommandEnum).includes(command as SvgCommandEnum)
-        ) {
-            throw new Error(`Invalid command: ${command}`);
-        }
-        return command as SvgCommandEnum;
-    }
-
-    static checkCoordinatesWithCommand(
-        command: SvgCommandEnum,
-        coordinates: Coordinate[],
-    ) {
-        switch (command) {
-            case SvgCommandEnum.MOVE:
-            case SvgCommandEnum.LINE: {
-                const expected = 1;
-                if (coordinates.length !== expected) {
-                    throw new Error(
-                        `Invalid number of coordinates for ${command} command. Expected ${expected}, got ${coordinates.length}`,
-                    );
-                }
-                break;
-            }
-            case SvgCommandEnum.QUADRATIC: {
-                const expected = 2;
-                if (coordinates.length !== expected) {
-                    throw new Error(
-                        `Invalid number of coordinates for ${command} command. Expected ${expected}, got ${coordinates.length}`,
-                    );
-                }
-                break;
-            }
-            case SvgCommandEnum.CUBIC: {
-                const expected = 3;
-                if (coordinates.length !== expected) {
-                    throw new Error(
-                        `Invalid number of coordinates for ${command} command. Expected ${expected}, got ${coordinates.length}`,
-                    );
-                }
-                break;
-            }
-            case SvgCommandEnum.CLOSE: {
-                const expected = 0;
-                if (coordinates.length !== 0) {
-                    throw new Error(
-                        `Invalid number of coordinates for ${command} command. Expected ${expected}, got ${coordinates.length}`,
-                    );
-                }
-                break;
-            }
-        }
-    }
-
-    /**
-     * Converts the ShapePoint to a string representation in SVG path format
-     * @returns A string in the format "command x, y "
-     */
-    toString() {
-        const array = this.toArray();
-        return `${this.command} ${array.slice(1).join(" ")} `;
-    }
-
-    /**
-     * Converts an array of ShapePoints into a single SVG path string
-     * @param points Array of ShapePoint objects to convert
-     * @returns Combined string of all points in SVG path format
-     */
-    static pointsToString(points: ShapePoint[]) {
-        let output = "";
-        for (const p of points) output += p.toString();
-        return output;
-    }
-
-    /**
-     * Converts an array of command-coordinate tuples into an array of ShapePoint objects.
-     * @param array - An array of command-coordinate tuples, where the first element is the command string and the remaining elements are the coordinate values.
-     * @returns An array of ShapePoint objects representing the input array.
-     */
-    static fromArray(array: VanillaPoint[]): ShapePoint[] {
-        const points: ShapePoint[] = [];
-        for (const point of array) {
-            const command = point[0] as string;
-            const coordinates = point.slice(1) as number[];
-            const coordPairs: Coordinate[] = [];
-            for (let i = 0; i < coordinates.length; i += 2) {
-                if (i + 1 >= coordinates.length) {
-                    console.warn(
-                        `Warning: Incomplete coordinate pair for command ${command}`,
-                    );
-                    break;
-                }
-                coordPairs.push({ x: coordinates[i], y: coordinates[i + 1] });
-            }
-
-            points.push(new ShapePoint(this.checkCommand(command), coordPairs));
-        }
-        return points;
-    }
-
-    /**
-     * Applies an offset to the coordinates of the ShapePoint in place.
-     *
-     * @param offset - An object containing the x and y offsets to apply.
-     * @returns The modified ShapePoint object (which is the same object).
-     */
-    applyOffset(offset: Coordinate) {
-        for (const point of this.coordinates) {
-            point.x += offset.x;
-            point.y += offset.y;
-        }
-        return this;
-    }
-
-    /**
-     * Converts the ShapePoint into an array format with command as first element
-     * @returns Tuple with command string followed by coordinate numbers
-     */
-    toArray(): VanillaPoint {
-        return [
-            this.command,
-            ...this.coordinates.flatMap((coord) => [coord.x, coord.y]),
-        ];
-    }
-
-    /**
-     * Converts an array of ShapePoints into an array of command-coordinate tuples
-     * @param points Array of ShapePoint objects to convert
-     * @returns Array of tuples, each containing a command string and coordinates
-     */
-    static pointsToArray(points: ShapePoint[]): VanillaPoint[] {
-        return points.map((point) => point.toArray());
-    }
-
-    /**
-     * Creates an array of `ShapePoint` objects from an SVG path string.
-     *
-     * This method takes advantage of the fact that the `fabric.Path` class can parse SVG paths.
-     *
-     * @param svgPath - The SVG path string to parse.
-     * @returns An array of `ShapePoint` objects representing the parsed path.
-     */
-    static fromString(svgPath: string): ShapePoint[] {
-        // Take advantage of the fact that fabric.Path can parse SVG paths
-        const tempPath = new fabric.Path(svgPath);
-        const points = ShapePoint.fromArray(
-            tempPath.path as any as VanillaPoint[],
-        );
-        return points;
-    }
-
-    /******************** GENERATORS *********************/
-    static createShapePoint(
-        command: SvgCommandEnum,
-        coordinates: Coordinate[],
-    ): ShapePoint {
-        this.checkCoordinatesWithCommand(command, coordinates);
-        return new ShapePoint(command, coordinates);
-    }
-
-    /**
-     * Creates a Move command ShapePoint
-     * @param x X coordinate to move to
-     * @param y Y coordinate to move to
-     * @returns New ShapePoint with "M" command
-     */
-    static Move(coordinate: Coordinate): ShapePoint {
-        return new ShapePoint(SvgCommandEnum.MOVE, [coordinate]);
-    }
-
-    /**
-     * Creates a Quadratic curve command ShapePoint
-     * @param cx Control point X coordinate
-     * @param cy Control point Y coordinate
-     * @param x End point X coordinate
-     * @param y End point Y coordinate
-     * @returns New ShapePoint with "Q" command
-     */
-    static Quadratic(
-        controlPoint: Coordinate,
-        endPoint: Coordinate,
-    ): ShapePoint {
-        return new ShapePoint(SvgCommandEnum.QUADRATIC, [
-            { x: controlPoint.x, y: controlPoint.y },
-            { x: endPoint.x, y: endPoint.y },
-        ]);
-    }
-
-    /**
-     * Creates a Cubic Bézier curve command ShapePoint.
-     * @param cx1 The X coordinate of the first control point.
-     * @param cy1 The Y coordinate of the first control point.
-     * @param cx2 The X coordinate of the second control point.
-     * @param cy2 The Y coordinate of the second control point.
-     * @param x The X coordinate of the end point.
-     * @param y The Y coordinate of the end point.
-     * @returns A new ShapePoint with the "C" command.
-     */
-    static Cubic(
-        controlPoint1: Coordinate,
-        controlPoint2: Coordinate,
-        endPoint: Coordinate,
-    ): ShapePoint {
-        return new ShapePoint(SvgCommandEnum.CUBIC, [
-            { x: controlPoint1.x, y: controlPoint1.y },
-            { x: controlPoint2.x, y: controlPoint2.y },
-            { x: endPoint.x, y: endPoint.y },
-        ]);
-    }
-
-    /**
-     * Creates a Line command ShapePoint.
-     * @param x The X coordinate of the end point.
-     * @param y The Y coordinate of the end point.
-     * @returns A new ShapePoint with the "L" command.
-     */
-    static Line(endPoint: Coordinate): ShapePoint {
-        return new ShapePoint(SvgCommandEnum.LINE, [endPoint]);
-    }
-
-    /**
-     * Creates a ShapePoint with the "Z" command, which closes the current path by drawing a straight line from the current point to the start point of the path.
-     * @returns A new ShapePoint with the "Z" command.
-     */
-    static Close(): ShapePoint {
-        return new ShapePoint(SvgCommandEnum.CLOSE, []);
-    }
-
-    // /**
-    //  * Creates an Arc command ShapePoint.
-    //  * @param rx The radius of the ellipse in the X-axis.
-    //  * @param ry The radius of the ellipse in the Y-axis.
-    //  * @param x The X coordinate of the end point.
-    //  * @param y The Y coordinate of the end point.
-    //  * @returns A new ShapePoint with the "A" command.
-    //  */
-    // static Arch(rx: number, ry: number, x: number, y: number): ShapePoint {
-    //     return new ShapePoint("A", [
-    //         { x: rx, y: ry },
-
-    //         { x, y },
-    //     ]);
-    // }
 }
 
 /**
