+++ apps/desktop/src/global/classes/canvasObjects/__test__/StaticMarcherShape.test.ts
@@ -1,8 +1,10 @@
 import { beforeEach, describe, expect, it } from "vitest";
-import { ShapePath, ShapePoint, VanillaPoint } from "../StaticMarcherShape";
 import OpenMarchCanvas from "../OpenMarchCanvas";
 import FieldPropertiesTemplates from "../../FieldProperties.templates";
 import { falsyUiSettings } from "@/components/canvas/__test__/MocksForCanvas";
+import { ShapePath } from "../ShapePath";
+import { ShapePoint } from "../ShapePoint";
+import { VanillaPoint } from "../StaticMarcherShape";
 
 describe.todo("StaticMarcherShape", () => {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
